+++ src/factory/action.ts
@@ -63,7 +63,7 @@ export type IAdditionalProperty = IPropertyValue<string>[];
  */
 export interface IAttributes<T extends ActionType, TObject, TResult> {
     identifier?: string;
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     /**
      * A property-value pair representing an additional characteristics of the entitity,
      * e.g. a product feature or another characteristic for which there is no matching property in schema.org.