+++ src/factory/event/screeningEvent.ts
@@ -109,7 +109,7 @@ export interface ISeller {
  * イベントに対するオファー
  */
 export interface IOffer {
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     typeOf: OfferType.Offer;
     priceCurrency: PriceCurrency.JPY;
     /**
@@ -137,6 +137,7 @@ export interface IOffer {
     unacceptedPaymentMethod?: string[];
     seller: ISeller;
 }
+export type IOffer4COA = Pick<IOffer, 'project' | 'typeOf' | 'offeredThrough' | 'priceCurrency'>;
 export import ITicketPriceComponent = OfferFactory.ITicketPriceComponent;
 export import ITicketPriceSpecification = OfferFactory.ITicketPriceSpecification;
 /**
@@ -217,7 +218,7 @@ export type ICOAOffer = COA.factory.reserve.IUpdReserveTicket & {
 };
 export type IWorkPerformed = ScreeningEventSeriesFactory.IWorkPerformed;
 export interface ILocation {
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     /**
      * 場所タイプ
      */
@@ -280,7 +281,7 @@ export interface IAttributes extends EventFactory.IAttributes<EventType.Screenin
     /**
      * 販売情報
      */
-    offers?: IOffer;
+    offers?: IOffer | IOffer4COA;
     /**
      * 発券数
      */