+++ src/factory/priceSpecification.ts
@@ -30,7 +30,7 @@ export interface IAccounting {
  * 価格仕様インターフェース
  */
 export interface IPriceSpecification<T extends PriceSpecificationType> {
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     id?: string;
     typeOf: T;
     name?: string | IMultilingualString;