+++ src/factory/task/placeOrder.ts
@@ -5,7 +5,7 @@ import * as TaskFactory from '../task';
 import { TaskName } from '../taskName';
 
 export interface IData {
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     object: { confirmationNumber: string; orderNumber: string };
     potentialActions?: IOrderActionPotentialActions;
 }