+++ src/factory/task/sendOrder.ts
@@ -5,7 +5,7 @@ import * as TaskFactory from '../task';
 import { TaskName } from '../taskName';
 
 export interface IData {
-    project: IProject;
+    project: Pick<IProject, 'id' | 'typeOf'>;
     object: SendOrderActionFactory.IObject;
     potentialActions?: SendOrderActionFactory.IPotentialActions;
 }